Add files via upload
The client says:
"My company, "Brooklyn Outdoor Film Festival" just got the approval to show movies at Brooklyn Bridge Park this year from August 5th through 8th, from 6pm to midnight. We can't wait to help the public discover awesome films!

I need to publicize the festival, and to do this, I need a website. On this site, I need to display information about the festival, the films we're showing, and regularly published news or announcements about the festival.

There should also be links to the city of Brooklyn’s news website, the Brooklyn Bridge Park’s website, and our Twitter, Instagram, and Facebook accounts somewhere on the page (navigation bar, footer links, etc). Since I haven't created these accounts yet, you can just have each social media icon as a link to its homepage (ex. https://twitter.com).

Shows will be free and open to everyone, but I want people to be able to pre-register so that I'm able to estimate how many people might come and so that we can close registration if it reaches the venue’s capacity of 5,000 people.

I'd like to have a professional web address, either a .com or .org -- you decide, but just make sure it's relevant to the film festival."

Project to-dos:
Website prototype: A website prototype created in HTML and CSS with Bootstrap.
Project brief: A project brief that covers logistical details about the overall project as well as technical details about the website.
This project includes:
A project brief explaining the client's needs, including an explanation of the technologies chosen.
A prototype folder with the HTML and CSS source code + images.